+++ packages/edge-bundler/node/manifest.ts
@@ -64,16 +64,24 @@ interface Manifest {
   function_config: Record<string, EdgeFunctionConfig>
 }
 
-interface GenerateManifestOptions {
-  bundles?: Bundle[]
-  declarations?: Declaration[]
-  featureFlags?: FeatureFlags
+interface GenerateManifestFunctionConfigOptions {
   functions: EdgeFunction[]
-  importMap?: string
   internalFunctionConfig?: Record<string, FunctionConfig>
-  layers?: Layer[]
   userFunctionConfig?: Record<string, FunctionConfig>
+}
+
+interface GenerateManifestRoutesOptions {
+  functions: EdgeFunction[]
+  declarations?: Declaration[]
+}
+
+interface GenerateManifestOptionsBase {
+  bundles?: Bundle[]
+  featureFlags?: FeatureFlags
+  importMap?: string
+  layers?: Layer[]
   bundlingTiming?: BundlingTiming
+  functions: EdgeFunction[]
 }
 
 const removeEmptyConfigValues = (functionConfig: EdgeFunctionConfig) =>
@@ -147,23 +155,14 @@ const normalizeMethods = (method: unknown, name: string): string[] | undefined =
   })
 }
 
-const generateManifest = ({
-  bundles = [],
-  declarations = [],
+export const generateManifestFunctionConfig = ({
   functions,
   userFunctionConfig = {},
   internalFunctionConfig = {},
-  importMap,
-  layers = [],
-  bundlingTiming,
-}: GenerateManifestOptions) => {
-  const preCacheRoutes: Route[] = []
-  const postCacheRoutes: Route[] = []
+}: GenerateManifestFunctionConfigOptions) => {
   const manifestFunctionConfig: Manifest['function_config'] = Object.fromEntries(
     functions.map(({ name }) => [name, { excluded_patterns: [] }]),
   )
-  const routedFunctions = new Set<string>()
-  const declarationsWithoutFunction = new Set<string>()
 
   for (const [name, singleUserFunctionConfig] of Object.entries(userFunctionConfig)) {
     // If the config block is for a function that is not defined, discard it.
@@ -210,6 +209,15 @@ const generateManifest = ({
     }
   }
 
+  return sanitizeEdgeFunctionConfig(manifestFunctionConfig)
+}
+
+export const generateManifestRoutes = ({ functions, declarations = [] }: GenerateManifestRoutesOptions) => {
+  const preCacheRoutes: Route[] = []
+  const postCacheRoutes: Route[] = []
+  const routedFunctions = new Set<string>()
+  const declarationsWithoutFunction = new Set<string>()
+
   declarations.forEach((declaration) => {
     const func = functions.find(({ name }) => declaration.function === name)
 
@@ -254,14 +262,53 @@ const generateManifest = ({
       preCacheRoutes.push(route)
     }
   })
+
+  const unroutedFunctions = functions.filter(({ name }) => !routedFunctions.has(name)).map(({ name }) => name)
+
+  return {
+    preCacheRoutes: preCacheRoutes.filter(nonNullable),
+    postCacheRoutes: postCacheRoutes.filter(nonNullable),
+    unroutedFunctions,
+    declarationsWithoutFunction,
+  }
+}
+
+type GenerateManifestOptions = GenerateManifestOptionsBase &
+  (
+    | GenerateManifestFunctionConfigOptions
+    | { manifestFunctionConfig: ReturnType<typeof generateManifestFunctionConfig> }
+  ) &
+  (GenerateManifestRoutesOptions | { manifestRoutes: ReturnType<typeof generateManifestRoutes> })
+
+const generateManifest = ({
+  bundles = [],
+  importMap,
+  layers = [],
+  bundlingTiming,
+  functions,
+  ...rest
+}: GenerateManifestOptions) => {
+  const manifestFunctionConfig =
+    'manifestFunctionConfig' in rest
+      ? rest.manifestFunctionConfig
+      : generateManifestFunctionConfig({ functions, ...rest })
+
+  const { preCacheRoutes, postCacheRoutes, unroutedFunctions, declarationsWithoutFunction } =
+    'manifestRoutes' in rest
+      ? rest.manifestRoutes
+      : generateManifestRoutes({
+          functions,
+          ...rest,
+        })
+
   const manifestBundles = bundles.map(({ extension, format, hash }) => ({
     asset: hash + extension,
     format,
   }))
   const manifest: Manifest = {
     bundles: manifestBundles,
-    routes: preCacheRoutes.filter(nonNullable),
-    post_cache_routes: postCacheRoutes.filter(nonNullable),
+    routes: preCacheRoutes,
+    post_cache_routes: postCacheRoutes,
     bundler_version: getPackageVersion(),
     layers,
     import_map: importMap,
@@ -270,7 +317,6 @@ const generateManifest = ({
       ? { bundling_timing: bundlingTiming }
       : {}),
   }
-  const unroutedFunctions = functions.filter(({ name }) => !routedFunctions.has(name)).map(({ name }) => name)
 
   return { declarationsWithoutFunction: [...declarationsWithoutFunction], manifest, unroutedFunctions }
 }
@@ -366,7 +412,7 @@ const getExcludedRegularExpressions = (declaration: Declaration): string[] => {
   return []
 }
 
-interface WriteManifestOptions extends GenerateManifestOptions {
+type WriteManifestOptions = GenerateManifestOptions & {
   distDirectory: string
 }