+++ packages/config/CHANGELOG.md
@@ -76,6 +76,29 @@
   * dependencies
     * netlify bumped from ^13.1.8 to ^13.1.9
 
+## [24.6.0](https://github.com/netlify/build/compare/config-v24.5.1...config-v24.6.0) (2026-05-21)
+
+
+### Features
+
+* add additional function config properties ([#7064](https://github.com/netlify/build/issues/7064)) ([41bf8da](https://github.com/netlify/build/commit/41bf8dace0b0b2e8ab55e6b85cdb314fed12a208))
+
+## [24.5.1](https://github.com/netlify/build/compare/config-v24.5.0...config-v24.5.1) (2026-05-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/api bumped from ^14.0.18 to ^14.0.19
+
+## [24.5.0](https://github.com/netlify/build/compare/config-v24.4.4...config-v24.5.0) (2026-04-14)
+
+
+### Features
+
+* rename `@netlify/db` to `@netlify/database` ([#7026](https://github.com/netlify/build/issues/7026)) ([3361c94](https://github.com/netlify/build/commit/3361c94f9572f2afed2cbb23b420b318a94674f7))
+
 ## [24.4.4](https://github.com/netlify/build/compare/config-v24.4.3...config-v24.4.4) (2026-03-20)